backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
anthropic==0.40.0
pydantic==2.10.3
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.1.3
matplotlib==3.9.3
scipy==1.14.1
yfinance==0.2.50
seaborn==0.13.2
requests==2.32.3
